% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{square_data}
\alias{square_data}
\title{Sample vegetation data with square grid}
\format{
An \code{sf} object with 300 rows and 6 variables.
}
\usage{
data(square_data)
}
\description{
A generated dataset containing presences and absences of five vegetation types.
}
\author{
Krisztina Dóra Konrád \email{konrad.krisztina@ecolres.hu}
}
\keyword{datasets}
